improved assert, added rcodeValue, made getCOunt return short
git-svn-id: http://svn.code.sf.net/p/dnsjava/code/trunk@54 c76caeb1-94fd-44dd-870f-0c9d92034fc1

+++ org/xbill/DNS/dnsHeader.java
@@ -109,7 +109,7 @@ decCount(int field) {
 	counts[field]--;
 }
 
-public int
+public short
 getCount(int field) {
 	return counts[field];
 }
